% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.test.R
\name{fun.test}
\alias{fun.test}
\title{A search Function}
\usage{
fun.test(seedwords, originalTextFile = originalTextFile, listname = NULL)
}
\arguments{
\item{x}{Any string.}
}
\description{
This function allows you to match the seedwords to specific article.
}
\examples{
fun.test()
}
\keyword{removeNumbers}
