% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeNumbers.R
\name{removeNumbers}
\alias{removeNumbers}
\title{A preprocess Function}
\usage{
removeNumbers(x)
}
\arguments{
\item{x}{Any string.}
}
\description{
This function allows you to remove special characters and number.
}
\examples{
removeNumbers()
}
\keyword{removeNumbers}
