% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmcmbc.R
\name{tmcmbc}
\alias{tmcmbc}
\title{A main Function}
\usage{
tmcmbc(localfreq = 0.005, ratio = 2)
}
\arguments{
\item{localfreq}{localfreq.}

\item{ratio}{localfreq/globalfreq}
}
\description{
This function allows you to understand articles.
}
\examples{
tmcmbc()
}
\keyword{tmbc}
